Changed type of coordinates from Floats to CGFloats

+++ sketch-pro/sketch-pro/BezierPointView.swift
@@ -11,8 +11,8 @@ import Cocoa
 class BezierPointView: NSView {
 	
 	var model : BezierPointMO!
-	var x : Float!
-	var y : Float!
+	var x : CGFloat!
+	var y : CGFloat!
 	var forwardControlPoint : ControlPointView!
 	var backwardControlPoint : ControlPointView!
 	var forwardControlPointExtension : ControlPointExtension!
